requirements.txt
openai==0.27.0
textract==1.6.3
python-docx==0.8.10
argparse==1.4.0
